badge_in/requirements-badge-in.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
python-dotenv
gpiozero
mfrc522-python
pyyaml
RPi.GPIO